Vagrant setup for pimometer project
This patch includes basic settings for pimometer to work in a vagrant
environment.  Henceforth, all one must do to get a working development
version of pimometer is "vagrant up" in the top level directory.  The
end result is a (libvirt) VM running with the required software
installed and services running.