Beginnings of REST api
This patch includes the ground work for the django driven rest api using
djangorestframework.  The vagrant VM now uses pip to install the
djangorestframework because the package was not available for the distro
of the box image.  An example, yet still non-functional, "User" and
"Group" api have been set in urls.py.  These will likely be something
more along the lines of "event/sensor1" or "event/sensor2" in the future
(might even include a username or auth token for the API).  Also added
some basic requirements for djangorestframework in settings.py to get it
basically functional.